# RoamOpenAPI/Sources/RoamOpenAPI/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Roam API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/reels": {
      "get": {
        "summary": "Listreels",
        "operationId": "listReels_api_reels_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SavedReelListItem"
                  },
                  "title": "Response Listreels Api Reels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reels/{reelId}": {
      "get": {
        "summary": "Getreel",
        "operationId": "getReel_api_reels__reelId__get",
        "parameters": [
          {
            "name": "reelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reelid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedReelDetailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reels/{reelId}/ideas": {
      "post": {
        "summary": "Createideaonreel",
        "operationId": "createIdeaOnReel_api_reels__reelId__ideas_post",
        "parameters": [
          {
            "name": "reelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reelid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIdeaOnReelBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdeaRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reels/{reelId}/promote": {
      "post": {
        "summary": "Promotereel",
        "operationId": "promoteReel_api_reels__reelId__promote_post",
        "parameters": [
          {
            "name": "reelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reelid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteReelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoteReelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateIdeaOnReelBody": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "notes": {
            "type": "string",
            "title": "Notes",
            "default": ""
          },
          "placeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Placeid"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "CreateIdeaOnReelBody"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdeaRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "title": "Userid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "notes": {
            "type": "string",
            "title": "Notes"
          },
          "sourceUrl": {
            "type": "string",
            "title": "Sourceurl"
          },
          "displayName": {
            "type": "string",
            "nullable": true,
            "title": "Displayname"
          },
          "placeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Placeid"
          },
          "savedReelId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Savedreelid"
          },
          "status": {
            "$ref": "#/components/schemas/IdeaStatus"
          },
          "planCount": {
            "type": "integer",
            "title": "Plancount",
            "default": 0
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "pipelineResult": {
            "$ref": "#/components/schemas/PipelineResultRead",
            "nullable": true
          },
          "placeName": {
            "type": "string",
            "nullable": true,
            "title": "Placename"
          },
          "placeLatitude": {
            "type": "number",
            "nullable": true,
            "title": "Placelatitude"
          },
          "placeLongitude": {
            "type": "number",
            "nullable": true,
            "title": "Placelongitude"
          }
        },
        "type": "object",
        "required": [
          "id",
          "userId",
          "title",
          "notes",
          "sourceUrl",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "IdeaRead"
      },
      "IdeaStatus": {
        "type": "string",
        "enum": [
          "captured",
          "suggesting",
          "ready",
          "planned"
        ],
        "title": "IdeaStatus"
      },
...
      "PipelineResultRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "ideaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Ideaid"
          },
          "jobId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Jobid"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "refinedTitle": {
            "type": "string",
            "nullable": true,
            "title": "Refinedtitle"
          },
          "category": {
            "type": "string",
            "nullable": true,
            "title": "Category"
          },
          "estimatedMinutes": {
            "type": "integer",
            "nullable": true,
            "title": "Estimatedminutes"
          },
          "modelName": {
            "type": "string",
            "nullable": true,
            "title": "Modelname"
          },
          "promptVersion": {
            "type": "string",
            "nullable": true,
            "title": "Promptversion"
          },
          "rawOutput": {
            "additionalProperties": true,
            "type": "object",
            "nullable": true,
            "title": "Rawoutput"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "placeSuggestions": {
            "items": {
              "$ref": "#/components/schemas/PlaceSuggestionRead"
            },
            "type": "array",
            "title": "Placesuggestions",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "createdAt"
        ],
        "title": "PipelineResultRead"
      },
...
      "PlaceSuggestionRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "resultId": {
            "type": "string",
            "format": "uuid",
            "title": "Resultid"
          },
          "placeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Placeid"
          },
          "rawName": {
            "type": "string",
            "nullable": true,
            "title": "Rawname"
          },
          "confidence": {
            "type": "number",
            "nullable": true,
            "title": "Confidence"
          },
          "isSelected": {
            "type": "boolean",
            "title": "Isselected"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "placeName": {
            "type": "string",
            "nullable": true,
            "title": "Placename"
          }
        },
        "type": "object",
        "required": [
          "id",
          "resultId",
          "isSelected",
          "createdAt"
        ],
        "title": "PlaceSuggestionRead"
      },
...
      "PromoteReelRequest": {
        "properties": {
          "promotions": {
            "items": {
              "$ref": "#/components/schemas/ReelPromotionItem"
            },
            "type": "array",
            "title": "Promotions"
          }
        },
        "type": "object",
        "title": "PromoteReelRequest"
      },
      "PromoteReelResponse": {
        "properties": {
          "ideaIds": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Ideaids"
          },
          "reelStatus": {
            "$ref": "#/components/schemas/SavedReelStatus"
          }
        },
        "type": "object",
        "required": [
          "reelStatus"
        ],
        "title": "PromoteReelResponse"
      },
      "ReelIngestCandidateRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "savedReelId": {
            "type": "string",
            "format": "uuid",
            "title": "Savedreelid"
          },
          "sortIndex": {
            "type": "integer",
            "title": "Sortindex"
          },
          "previewTitle": {
            "type": "string",
            "title": "Previewtitle",
            "default": ""
          },
          "isSynthetic": {
            "type": "boolean",
            "title": "Issynthetic",
            "default": false
          },
          "resolvedPlaceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Resolvedplaceid"
          },
          "promotedIdeaId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Promotedideaid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "resolvedPlaceName": {
            "type": "string",
            "nullable": true,
            "title": "Resolvedplacename"
          }
        },
        "type": "object",
        "required": [
          "id",
          "savedReelId",
          "sortIndex",
          "createdAt"
        ],
        "title": "ReelIngestCandidateRead"
      },
      "ReelPromotionItem": {
        "properties": {
          "candidateId": {
            "type": "string",
            "format": "uuid",
            "title": "Candidateid"
          },
          "title": {
            "type": "string",
            "nullable": true,
            "title": "Title"
          },
          "mapsQuery": {
            "type": "string",
            "nullable": true,
            "title": "Mapsquery"
          },
          "placeAddress": {
            "type": "string",
            "nullable": true,
            "title": "Placeaddress"
          },
          "category": {
            "type": "string",
            "nullable": true,
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "candidateId"
        ],
        "title": "ReelPromotionItem"
      },
...
      "SavedReelDetailRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "jobId": {
            "type": "string",
            "format": "uuid",
            "title": "Jobid"
          },
          "reelUrl": {
            "type": "string",
            "title": "Reelurl"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/SavedReelStatus"
          },
          "thumbnailSignedUrl": {
            "type": "string",
            "nullable": true,
            "title": "Thumbnailsignedurl"
          },
          "jobStatus": {
            "type": "string",
            "title": "Jobstatus"
          },
          "jobError": {
            "type": "string",
            "nullable": true,
            "title": "Joberror"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ReelIngestCandidateRead"
            },
            "type": "array",
            "title": "Candidates"
          },
          "ideas": {
            "items": {
              "$ref": "#/components/schemas/SavedReelIdeaSummary"
            },
            "type": "array",
            "title": "Ideas"
          },
          "ideaIds": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Ideaids"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "jobId",
          "reelUrl",
          "title",
          "status",
          "jobStatus",
          "createdAt",
          "updatedAt"
        ],
        "title": "SavedReelDetailRead"
      },
      "SavedReelIdeaSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/IdeaStatus"
          },
          "placeId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "title": "Placeid"
          },
          "placeName": {
            "type": "string",
            "nullable": true,
            "title": "Placename"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "status"
        ],
        "title": "SavedReelIdeaSummary"
      },
      "SavedReelListItem": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "jobId": {
            "type": "string",
            "format": "uuid",
            "title": "Jobid"
          },
          "reelUrl": {
            "type": "string",
            "title": "Reelurl"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/SavedReelStatus"
          },
          "thumbnailSignedUrl": {
            "type": "string",
            "nullable": true,
            "title": "Thumbnailsignedurl"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "jobId",
          "reelUrl",
          "title",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "title": "SavedReelListItem"
      },
      "SavedReelStatus": {
        "type": "string",
        "enum": [
          "processing",
          "needs_review",
          "promoted",
          "failed"
        ],
        "title": "SavedReelStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}